+++ packages/block-editor/src/components/block-card/index.js
@@ -7,7 +7,11 @@ import clsx from 'clsx';
  * WordPress dependencies
  */
 import deprecated from '@wordpress/deprecated';
-import { Button } from '@wordpress/components';
+import {
+	Button,
+	__experimentalText as Text,
+	__experimentalVStack as VStack,
+} from '@wordpress/components';
 import { chevronLeft, chevronRight } from '@wordpress/icons';
 import { __, isRTL } from '@wordpress/i18n';
 import { useSelect, useDispatch } from '@wordpress/data';
@@ -60,14 +64,14 @@ function BlockCard( { title, icon, description, blockType, className } ) {
 				/>
 			) }
 			<BlockIcon icon={ icon } showColors />
-			<div className="block-editor-block-card__content">
+			<VStack spacing={ 1 }>
 				<h2 className="block-editor-block-card__title">{ title }</h2>
 				{ description && (
-					<span className="block-editor-block-card__description">
+					<Text className="block-editor-block-card__description">
 						{ description }
-					</span>
+					</Text>
 				) }
-			</div>
+			</VStack>
 		</div>
 	);
 }