+++ packages/dataviews/src/view-table.tsx
@@ -15,7 +15,6 @@ import {
 	privateApis as componentsPrivateApis,
 	CheckboxControl,
 	Spinner,
-	VisuallyHidden,
 } from '@wordpress/components';
 import {
 	forwardRef,
@@ -549,9 +548,9 @@ function ViewTable< Item extends AnyItem >( {
 								data-field-id="actions"
 								className="dataviews-view-table__actions-column"
 							>
-								<VisuallyHidden>
+								<span className="dataviews-view-table-header">
 									{ __( 'Actions' ) }
-								</VisuallyHidden>
+								</span>
 							</th>
 						) }
 					</tr>